+++ src/node/ext/ldap/ugm/__init__.py
@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-from _api import (
+from ._api import (
     PrincipalsConfig,
     UsersConfig,
     GroupsConfig,